+++ src/l1j/server/server/utils/CalcStat.java
@@ -255,7 +255,7 @@ public class CalcStat {
 			}
 		}
 		else if (charType == 1) { // 骑士
-			randommp = (int) (randommp * 2 / 3);
+			randommp = (randommp * 2 / 3);
 			if (baseMaxMp + randommp > Config.KNIGHT_MAX_MP) {
 				randommp = Config.KNIGHT_MAX_MP - baseMaxMp;
 			}
@@ -282,14 +282,14 @@ public class CalcStat {
 			}
 		}
 		else if (charType == 5) { // 龙骑士
-			randommp = (int) (randommp * 2 / 3);
+			randommp = (randommp * 2 / 3);
 
 			if (baseMaxMp + randommp > Config.DRAGONKNIGHT_MAX_MP) {
 				randommp = Config.DRAGONKNIGHT_MAX_MP - baseMaxMp;
 			}
 		}
 		else if (charType == 6) { // 幻术师
-			randommp = (int) (randommp * 5 / 3);
+			randommp = (randommp * 5 / 3);
 
 			if (baseMaxMp + randommp > Config.ILLUSIONIST_MAX_MP) {
 				randommp = Config.ILLUSIONIST_MAX_MP - baseMaxMp;
